+++ lib/rust/ensogl/component/grid-view/src/entry.rs
@@ -14,7 +14,13 @@ use ensogl_core::display::geometry::compound::sprite;
 use ensogl_core::display::scene::Layer;
 use ensogl_core::display::Attribute;
 
+
+// ==============
+// === Export ===
+// ==============
+
 pub mod visible;
+
 
 
 // ===============
