+++ lib/rust/ensogl/component/grid-view/src/selectable/highlight/shape.rs
@@ -25,6 +25,7 @@ use ensogl_core::data::color;
 use ensogl_scroll_area::Viewport;
 
 
+
 // ========================
 // === Shape Definition ===
 // ========================
